+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 ## Added
 
 - Fixed #892, adding login links to error messages prompting login.
+- Translate revised strings that are machine translated in other locales.
 
 ## Fixed